requirements.txt
asn1crypto==1.4.0
astroid==2.8.6
atomicwrites==1.4.0
attrs==21.2.0
azure-common==1.1.27
azure-core==1.20.1
azure-storage-blob==12.9.0
boto3==1.18.55
boto3-stubs==1.20.11
botocore==1.21.65
botocore-stubs==1.23.12
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==3.4.8
idna==3.3
iniconfig==1.1.1
isodate==0.6.0
isort==5.10.1
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
msrest==0.6.21
numpy==1.21.4
oauthlib==3.1.1
oscrypto==1.2.1
packaging==21.3
pandas==1.3.4
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pycryptodomex==3.11.0
PyJWT==2.3.0
pylint==2.11.1
pymongo==3.12.1
pyodbc==4.0.32
pyOpenSSL==20.0.1
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
snowflake-connector-python==2.6.2
toml==0.10.2
typing_extensions==4.0.0
urllib3==1.26.7
wrapt==1.13.3
